% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caches.R
\name{clear_resteasy_cache}
\alias{clear_resteasy_cache}
\alias{file_cache}
\alias{environment_cache}
\alias{null_cache}
\alias{is.cache}
\alias{as.cache}
\alias{cache_full}
\alias{set_cached_text}
\alias{get_cached_text}
\alias{set_cached_raw}
\alias{get_cached_raw}
\alias{clear_cache}
\alias{cache_size}
\alias{get_cached_text.file_cache}
\alias{get_cached_raw.file_cache}
\alias{set_cached_text.file_cache}
\alias{set_cached_raw.file_cache}
\alias{clear_cache.file_cache}
\alias{cache_size.file_cache}
\alias{get_cached_text.environment_cache}
\alias{get_cached_raw.environment_cache}
\alias{set_cached_text.environment_cache}
\alias{set_cached_raw.environment_cache}
\alias{cache_size.environment_cache}
\alias{clear_cache.environment_cache}
\alias{set_cached_text.null_cache}
\alias{set_cached_raw.null_cache}
\alias{get_cached_text.null_cache}
\alias{get_cached_raw.null_cache}
\alias{cache_size.null_cache}
\title{Clear cached results}
\usage{
clear_resteasy_cache(cache = NA)

file_cache(directory, max_size = Inf, ...)

environment_cache(env, max_size = 10000, ...)

null_cache()

is.cache(x)

as.cache(x, ...)

cache_full(cache)

set_cached_text(cache, url, data, ...)

get_cached_text(cache, url, ...)

set_cached_raw(cache, url, data, ...)

get_cached_raw(cache, url, ...)

clear_cache(cache)

cache_size(cache)

\method{get_cached_text}{file_cache}(cache, url, ...)

\method{get_cached_raw}{file_cache}(cache, url, ...)

\method{set_cached_text}{file_cache}(cache, url, data, ...)

\method{set_cached_raw}{file_cache}(cache, url, data, ...)

\method{clear_cache}{file_cache}(cache)

\method{cache_size}{file_cache}(cache)

\method{get_cached_text}{environment_cache}(cache, url, ...)

\method{get_cached_raw}{environment_cache}(cache, url, ...)

\method{set_cached_text}{environment_cache}(cache, url, data, ...)

\method{set_cached_raw}{environment_cache}(cache, url, data, ...)

\method{cache_size}{environment_cache}(cache)

\method{clear_cache}{environment_cache}(cache)

\method{set_cached_text}{null_cache}(cache, url, data, ...)

\method{set_cached_raw}{null_cache}(cache, url, data, ...)

\method{get_cached_text}{null_cache}(cache, url, ...)

\method{get_cached_raw}{null_cache}(cache, url, ...)

\method{cache_size}{null_cache}(cache)
}
\arguments{
\item{cache, x}{An environment, a directory name, or NA to clear
the default internal cache}

\item{directory}{a directory in which files wil be kept}

\item{max_size}{The maximum number of entries to keep in the cache}

\item{...}{passed to/from methods}

\item{env}{an environment}

\item{url}{An object whose digest is identical for identical requests}

\item{data}{the payload to cache}
}
\description{
Clears the local cache of downloaded files (by default, an
environment in the package namespace). Clearing a directory
cache will result in all files with the extention ".cached"
being deleted from that directory.
}
\examples{
clear_resteasy_cache()

}
