% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/easy_create.R
\name{easy_create}
\alias{easy_create}
\title{Create a file}
\usage{
easy_create(x, repo, path, message, branch = "master", ...)
}
\arguments{
\item{x}{A data.frame}

\item{repo}{(character) Repository name. Required.}

\item{path}{(character) Path to file from root of repo. Required.}

\item{message}{(character) Commmit message. Required.}

\item{branch}{(character) Branch to use, Default: master}

\item{...}{Curl options passed on to \code{\link[httr]{PUT}}}
}
\description{
Create a file
}
\examples{
\dontrun{
options(github_name = "Scott Chamberlain")
options(github_email = "myrmecocystus@gmail.com")
easy_create(x=iris, repo = "sckott/testeasy", path ="iris.csv",
   message = "hello world")
}
}

