% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/easy_delete.R
\name{easy_delete}
\alias{easy_delete}
\title{Delete a file}
\usage{
easy_delete(repo, path, message, branch = "master", ...)
}
\arguments{
\item{repo}{(character) Repository name. Required.}

\item{path}{(character) Path to file from root of repo. Required.}

\item{message}{(character) Commmit message. Required.}

\item{branch}{(character) Branch to use, Default: master}

\item{...}{Curl options passed on to \code{\link[httr]{DELETE}}}
}
\description{
Delete a file
}
\examples{
\dontrun{
options(github_name = "Scott Chamberlain")
options(github_email = "myrmecocystus@gmail.com")
easy_create(x=iris, repo = "ropensci/datasets", path ="iris.csv",
   message = "hello world")
easy_delete(repo = "ropensci/datasets", path ="iris.csv",
   message = "delete it!")
}
}

