% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/easy_get.R
\name{easy_get}
\alias{easy_get}
\title{Get a file with query parameters}
\usage{
easy_get(repo, path, branch = "master", curlopts = list(), ...)
}
\arguments{
\item{repo}{(character) Repository name}

\item{path}{(character) Path to file from root of repo}

\item{branch}{(character) Branch to use, Default: master}

\item{curlopts}{Curl options passed on to \code{\link[httr]{GET}}}

\item{...}{Further args to sort or filter}
}
\description{
Get a file with query parameters
}
\details{
You can sort and filter by column names. Pass in key/value pairs
to either sort of filter on.
}
\examples{
\dontrun{
rest_GET(repo = "datasets/gdp", path ="data/gdp.csv")
rest_GET("datasets/gdp", "data/gdp.csv", Year = 2000)

rest_GET(repo="ropensci/datasets", "mammals/mammals.csv")
easy_get(repo="ropensci/datasets", path="mammals/mammals.csv")
easy_get("ropensci/datasets", "mammals/mammals.csv", Family="Coryphodontidae")
easy_get("ropensci/datasets", "mammals/mammals.csv", Last_appearance_mya=50)
}
}

