% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/easy_readme.R
\name{easy_readme}
\alias{easy_readme}
\title{Get README file for a repo}
\usage{
easy_readme(repo, branch = "master", ...)
}
\arguments{
\item{repo}{(character) Repository name}

\item{branch}{(character) Git branch. Default: master}

\item{...}{Further args to \code{\link[httr]{GET}}}
}
\description{
Get README file for a repo
}
\examples{
\dontrun{
easy_readme(repo="sckott/testeasy")
}
}

