% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/easy_update.R
\name{easy_update}
\alias{easy_update}
\title{Update a file}
\usage{
easy_update(x, repo, path, message, branch = "master", ...)
}
\arguments{
\item{x}{A data.frame}

\item{repo}{(character) Repository name. Required.}

\item{path}{(character) Path to file from root of repo. Required.}

\item{message}{(character) Commmit message. Required.}

\item{branch}{(character) Branch to use, Default: master}

\item{...}{Curl options passed on to \code{\link[httr]{PUT}}}
}
\description{
Update a file
}
\examples{
\dontrun{
options(github_name = "Scott Chamberlain")
options(github_email = "myrmecocystus@gmail.com")
easy_create(iris, repo = "sckott/testeasy", path = "iris.csv",
   message = "hello world")
easy_update(iris[1:5,], repo = "sckott/testeasy", path = "iris.csv",
   message = "changes to iris file")
}
}

