% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_query.R
\name{get_text}
\alias{get_text}
\alias{get_textf}
\alias{get_raw}
\alias{get_rawf}
\alias{get_text_}
\alias{get_text_f}
\alias{get_raw_}
\alias{get_raw_f}
\alias{get_download}
\alias{get_downloadf}
\alias{get_download_}
\alias{get_download_f}
\title{Query an endpoint and cache the data}
\usage{
get_text(.endpoint, ..., .cache = NULL, .parser = identity,
  .quiet = FALSE, .encoding = NULL)

get_textf(.endpoint, ...)

get_raw(.endpoint, ..., .cache = NULL, .parser = identity, .quiet = FALSE)

get_rawf(.endpoint, ...)

get_text_(endpoint, params = list(), cache = NULL, parser = identity,
  quiet = FALSE, encoding = NULL, ...)

get_text_f(endpoint, ...)

get_raw_(endpoint, params = list(), cache = NULL, parser = identity,
  quiet = FALSE, ...)

get_raw_f(endpoint, ...)

get_download(.endpoint, ..., .cache = NULL, .parser = identity,
  .quiet = FALSE)

get_downloadf(.endpoint, ...)

get_download_(endpoint, params = list(), cache = NULL, parser = identity,
  quiet = FALSE, ...)

get_download_f(endpoint, ...)
}
\arguments{
\item{..., params}{Key/value pairs to send to the query (non url-encoded)}

\item{endpoint, .endpoint}{The URL endpoint to send the query}

\item{cache, .cache}{An environment, directory name, NULL (disable cache), or NA for internal cache}

\item{parser, .parser}{A function that the (character) results will be passed through}

\item{quiet, .quiet}{Use quiet=TRUE to supress messaging}

\item{encoding, .encoding}{The encoding to use (defaults to auto, see httr::\link[httr]{content})}
}
\value{
The result of parser on the text resulting from the call
}
\description{
Query an endpoint and cache the data
}
\examples{
# that's some catch, that catch-22:
get_text("https://www.goodreads.com/book/title",
         key="HSkIMuOGlxFIOmfBCGFVA", title='catch-22')
}
