% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/github_auth.R
\name{github_pat}
\alias{github_pat}
\title{Github auth}
\usage{
github_pat()
}
\arguments{
\item{appname}{Application name}

\item{key}{Application key}

\item{secret}{App secret key}
}
\description{
Github auth
}
\examples{
\dontrun{
github_auth()
}
}

