% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rate_limit.R
\name{rate_limit}
\alias{rate_limit}
\title{Rate limit}
\usage{
rate_limit(...)
}
\arguments{
\item{...}{Further args to \code{\link[httr]{GET}}}
}
\description{
Rate limit
}
\examples{
\dontrun{
rate_limit()
}
}

