% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rest_DELETE.R
\name{rest_DELETE}
\alias{rest_DELETE}
\title{Delete a file}
\usage{
rest_DELETE(repo, path, message, sha = NULL, branch = "master", ...)
}
\arguments{
\item{repo}{(character) Repository name}

\item{path}{(character) Path to file from root of repo}

\item{branch}{(character) Git branch. Default: master}

\item{...}{Further args to \code{\link[httr]{DELETE}}}
}
\description{
Delete a file
}
\examples{
\dontrun{
rest_DELETE(repo = "ropensci/datasets", path = "mtcars.csv",
   message = "deleting file")
}
}

