% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rest_GET.R
\name{rest_GET}
\alias{rest_GET}
\title{Get a file}
\usage{
rest_GET(repo, path, branch = "master", ...)
}
\arguments{
\item{repo}{(character) Repository name}

\item{path}{(character) Path to file from root of repo}

\item{branch}{(character) Git branch. Default: master}

\item{...}{Further args to \code{\link[httr]{GET}}}
}
\description{
Get a file
}
\examples{
\dontrun{
rest_GET(repo="ropensci/datasets", path="planttraits/morphological.csv")
rest_GET(repo="ropensci/datasets", path="mammals/mammals.csv")
rest_GET(repo="datasets/global-temp", path="data/annual.csv")
rest_GET("datasets/gold-prices", "data/data.csv")
rest_GET("datasets/gdp", "data/gdp.csv")
}
}

