% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rest_PUT.R
\name{rest_PUT}
\alias{rest_PUT}
\title{Create a file}
\usage{
rest_PUT(x, repo, path, message, branch = "master", sha = NULL, ...)
}
\arguments{
\item{repo}{(character) Repository name}

\item{path}{(character) Path to file from root of repo}

\item{message}{(character) Commmit message}

\item{branch}{(character) Branch to use}

\item{sha}{(character) The blob SHA of the file being replaced}

\item{...}{Curl options passed on to \code{\link[httr]{PUT}}}
}
\description{
Create a file
}
\examples{
\dontrun{
options(github_name = "Scott Chamberlain")
options(github_email = "myrmecocystus@gmail.com")
rest_PUT(x=mtcars, repo = "ropensci/datasets", path ="mtcars.csv",
   message = "hello world")
}
}

