% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restquery.R
\name{restquery}
\alias{restquery}
\title{Query an endpoint and cache the data}
\usage{
restquery(.endpoint, ..., .cache = ".api_result", .parser = NULL,
  .quiet = FALSE)
}
\arguments{
\item{.endpoint}{The URL endpoint to send the query}

\item{...}{Key/value pairs to send to the query (will be passed through URLEncode)}

\item{.cache}{A variable name in .GlobalEnv to use as a cache (NULL to disable)}

\item{.parser}{A function that the (character) results will be passed through}

\item{.quiet}{Use .quiet=TRUE to supress messaging}
}
\value{
The query result
}
\description{
Query an endpoint and cache the data
}
\examples{
restquery(.endpoint="https://www.goodreads.com/book/title",
          key="HSkIMuOGlxFIOmfBCGFVA", title='catch-22')

}

